backend/requirements.txt
alembic==1.15.2
greenlet==3.1.1
Mako==1.3.10
MarkupSafe==3.0.2
SQLAlchemy==2.0.40
typing_extensions==4.13.2
numpy>=1.24.3
scikit-learn>=1.3.0
fastapi>=0.68.1
uvicorn>=0.15.0
pydantic>=1.8.2
python-dotenv>=0.19.0 

